Thesis_Web/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pdfminer.six
pdf2image
pytesseract
joblib
Pillow
numpy
opencv-python-headless
python-multipart